+++ da-project-backend/app/types.py
@@ -1,6 +1,7 @@
 from enum import StrEnum
 
 
+## MODEL TYPES
 class PageChartType(StrEnum):
     BAR_CHART = "BAR_CHART"
     PIE_CHART = "PIE_CHART"
@@ -12,3 +13,15 @@ class ColumnDataType(StrEnum):
     BOOLEAN = "BOOLEAN"
     NUMBER = "NUMBER"
     STRING = "STRING"
+
+
+## API TYPES
+class ColumnOperation(StrEnum):
+    FIRST = "first"
+    LAST = "last"
+    MAX = "max"
+    MEAN = "mean"
+    MEDIAN = "median"
+    MIN = "min"
+    MODE = "mode"
+    SUM = "sum"
